Some Unit tests added.

+++ src/main/java/com/example/demo/repository/H2Repository.java
@@ -3,5 +3,5 @@ package com.example.demo.repository;
 import com.example.demo.model.H2Entity;
 import org.springframework.data.repository.CrudRepository;
 
-public interface H2Repository extends CrudRepository<H2Entity, Long> {
+public interface H2Repository extends CrudRepository<H2Entity, Integer> {
 }
